+++ inegraphics/src/main/java/com/inepex/inegraphics/shared/gobjects/PathElement.java
@@ -2,10 +2,10 @@ package com.inepex.inegraphics.shared.gobjects;
 
 public abstract class PathElement {
 
-	protected int endPointX, endPointY;
+	protected double endPointX, endPointY;
 	
 
-	protected PathElement( int endPointX,int endPointY) {
+	protected PathElement( double endPointX,double endPointY) {
 		
 		this.endPointX = endPointX;
 		this.endPointY = endPointY;
@@ -14,28 +14,28 @@ public abstract class PathElement {
 	/**
 	 * @return the endPointX
 	 */
-	public int getEndPointX() {
+	public double getEndPointX() {
 		return endPointX;
 	}
 
 	/**
 	 * @param endPointX the endPointX to set
 	 */
-	public void setEndPointX(int endPointX) {
+	public void setEndPointX(double endPointX) {
 		this.endPointX = endPointX;
 	}
 
 	/**
 	 * @return the endPointY
 	 */
-	public int getEndPointY() {
+	public double getEndPointY() {
 		return endPointY;
 	}
 
 	/**
 	 * @param endPointY the endPointY to set
 	 */
-	public void setEndPointY(int endPointY) {
+	public void setEndPointY(double endPointY) {
 		this.endPointY = endPointY;
 	}
